Nerfed Footpad
The way it was, it simply gave way too much Dodging for the survivor as he runs, he basically becomes the Flash in terms of dodging. The nerf makes it useful, just not too useful, and buff_duration = 1 stops the Dodge super train as soon as you stop or attack.